refactor: make settings_reminder_view settings_notifications_view
